# apps/backend/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Monaco API
  version: 0.1.0
  description: HTTP contract between the Monaco iOS app and the Monaco backend.
  contact:
    name: Monaco
    url: https://github.com/Munyon-Canyon/monaco
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:8080
    description: Local development server.
security:
  - bearerAuth: []
...
paths:
...
  /v1/me/referral-code:
    get:
      operationId: getMyReferralCode
      tags: [referrals]
      summary: Read the caller's invite code and links.
      description: >-
        Returns the caller's random code and its link. `handle_link` is set and `handle_unlocked` is
        true once the caller's first deposit is credited. A user whose code is not minted yet gets
        referral_code_pending, a retryable 503, so the app tries again.
      responses:
        "200":
          description: The caller's invite code.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyReferralCode"
        default:
          $ref: "#/components/responses/Problem"
  # api/spec/social.yaml
...
components:
...
  securitySchemes:
  # api/spec/base.yaml
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        A bearer token in the Authorization header. Operations that declare an empty security
        list, such as /healthz, take no token. The token format belongs to the identity module;
        local development and tests use HS256 dev tokens from `monacoctl dev token`.
  responses:
...
    Problem:
      description: An RFC 9457 problem. Switch on `code`; show `message`.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Problem:
      type: object
      description: An RFC 9457 problem details body with a stable Monaco error code.
      examples:
        - type: about:blank
          title: Not Found
          status: 404
          code: not_found
          message: Not found.
          trace_id: 4bf92f3577b34da6a3ce929d0e0e4736
          retryable: false
      additionalProperties: false
      required: [type, title, status, code, message, trace_id, retryable]
      properties:
        type:
          type: string
          description: Always about:blank. The code field carries the problem type.
          const: about:blank
        title:
          type: string
          description: The HTTP status phrase.
          examples: [Not Found]
        status:
          type: integer
          description: The HTTP status code of this response.
          minimum: 400
          maximum: 599
          examples: [404]
        code:
          $ref: "#/components/schemas/ErrorCode"
          examples: [not_found]
        message:
          type: string
          description: A user-facing sentence the app shows as is.
          examples: [Not found.]
        trace_id:
          type: string
          description: The W3C trace id of the request. Logs and traces carry the same id.
          pattern: "^[0-9a-f]{32}$"
          examples: [4bf92f3577b34da6a3ce929d0e0e4736]
        retryable:
          type: boolean
          description: True when the same request may succeed if sent again later.
          examples: [false]
  # api/spec/cabal.yaml
...
    ErrorCode:
      type: string
      description: The closed list of error codes. Generated from the errs table by cmd/gen errors.
      examples: [not_found]
      enum:
        - access_request_not_pending
        - account_banned
        - account_deleted
        - account_has_balance
        - account_has_positions
        - account_status_transition
        - account_suspended
        - already_member
        - analytics_pii
        - apns_auth_failed
        - apns_unavailable
        - asset_not_found
        - asset_untradable
        - auth_state_transition
        - cabal_banned
        - cabal_not_found
        - cabal_paused
        - calendar_expired
        - cannot_follow_self
        - cannot_revoke_access
        - client_closed
        - conservation_broken
        - db_schema_behind
        - db_unavailable
        - decode_failed
        - display_name_invalid
        - faultpoint
        - feed_item_not_found
        - forbidden
        - handle_invalid
        - handle_required
        - handle_reserved
        - handle_taken
        - handle_too_soon
        - idempotency_in_flight
        - idempotency_mismatch
        - insufficient_funds
        - internal
        - invalid_address
        - invalid_config
        - invalid_input
        - invite_expired
        - join_needs_request
        - jupiter_rejected
        - jupiter_unavailable
        - leave_creator_with_members
        - leave_holds_shares
        - leave_last_member_pot_not_empty
        - ledger_unbalanced
        - live_swap_exists
        - login_method_not_allowed
        - no_route
        - not_a_voter
        - not_cabal_creator
        - not_cabal_member
        - not_found
        - not_proposer
        - panic
        - phone_not_linked
        - photo_invalid
        - post_hog_rejected
        - post_hog_unavailable
        - pot_exceeded
        - pot_value_zero
        - price_unavailable
        - privy_unavailable
        - proposal_closed
        - proposal_not_found
        - proposal_still_open
        - rate_limited
        - referral_code_pending
        - referral_code_unknown
        - relayer_underfunded
        - request_not_needed
        - request_pending
        - rpc_unavailable
        - session_required
        - slippage_exceeded
        - storage_unavailable
        - swap_failed
        - swap_not_found
        - swap_not_retryable
        - unauthorized
        - upstream_timeout
        - upstream_unavailable
        - user_banned
        - user_not_found
        - version_conflict
        - wallet_mismatch
        - withdraw_not_allowed
        - x_not_linked
  # api/spec/funding.yaml
...
    MyReferralCode:
      type: object
      description: The caller's invite code and the links that carry it.
      examples:
        - code: k7m4qx2p
          link: https://monacolabs.xyz/r/k7m4qx2p
          handle_link: https://monacolabs.xyz/r/kaicenat
          handle_unlocked: true
      additionalProperties: false
      required: [code, link, handle_link, handle_unlocked]
      properties:
        code:
          type: string
          description: The random 8-character code. It never changes.
          pattern: "^[23456789abcdefghjkmnpqrstuvwxyz]{8}$"
          examples: [k7m4qx2p]
        link:
          type: string
          description: The invite link for the random code.
          examples: ["https://monacolabs.xyz/r/k7m4qx2p"]
        handle_link:
          type: [string, "null"]
          description: The invite link for the caller's handle, or null until the first deposit unlocks it.
          examples: ["https://monacolabs.xyz/r/kaicenat"]
        handle_unlocked:
          type: boolean
          description: True once the caller's first deposit is credited.
          examples: [true]
  # api/spec/social.yaml
...